requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pymysql>=1.1.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic[email]>=2.5.0
email-validator>=2.0.0

# RAG Chatbot
google-genai>=0.1.0
numpy>=1.26.0
openai>=1.0.0

